aemet_plugin/requirements.txt
aiohttp
pandas
python-dateutil